requirements.txt
google-generativeai>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
